Rename landing-page template to marketing-page
