+++ docs/data/material/getting-started/templates/marketing-page/MarketingPage.js
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { PaletteMode } from '@mui/material';
+import PropTypes from 'prop-types';
+
 import CssBaseline from '@mui/material/CssBaseline';
 import Box from '@mui/material/Box';
 import Divider from '@mui/material/Divider';
@@ -16,17 +17,9 @@ import Features from './components/Features';
 import Testimonials from './components/Testimonials';
 import FAQ from './components/FAQ';
 import Footer from './components/Footer';
-import getLPTheme from './getLPTheme';
+import getLPTheme from './getMPTheme';
 
-interface ToggleCustomThemeProps {
-  showCustomTheme: Boolean;
-  toggleCustomTheme: () => void;
-}
-
-function ToggleCustomTheme({
-  showCustomTheme,
-  toggleCustomTheme,
-}: ToggleCustomThemeProps) {
+function ToggleCustomTheme({ showCustomTheme, toggleCustomTheme }) {
   return (
     <Box
       sx={{
@@ -63,8 +56,15 @@ function ToggleCustomTheme({
   );
 }
 
-export default function LandingPage() {
-  const [mode, setMode] = React.useState<PaletteMode>('light');
+ToggleCustomTheme.propTypes = {
+  showCustomTheme: PropTypes.shape({
+    valueOf: PropTypes.func.isRequired,
+  }).isRequired,
+  toggleCustomTheme: PropTypes.func.isRequired,
+};
+
+export default function MarketingPage() {
+  const [mode, setMode] = React.useState('light');
   const [showCustomTheme, setShowCustomTheme] = React.useState(true);
   const LPtheme = createTheme(getLPTheme(mode));
   const defaultTheme = createTheme({ palette: { mode } });
