+++ docs/data/material/getting-started/templates/marketing-page/README.md
@@ -1,4 +1,4 @@
-# Landing page template
+# Marketing page template
 
 ## Usage
 
@@ -6,10 +6,10 @@
 
 1. Copy the files into your project, or one of the [example projects](https://github.com/mui/material-ui/tree/next/examples).
 2. Make sure your project has the required dependencies: @mui/material, @mui/icons-material, @emotion/styled, @emotion/react.
-3. Import and use the `LandingPage` component.
+3. Import and use the `MarketingPage` component.
 
 ## Demo
 
 <!-- #default-branch-switch -->
 
-View the demo at https://next.mui.com/material-ui/getting-started/templates/landing-page/.
+View the demo at https://next.mui.com/material-ui/getting-started/templates/marketing-page/.
