+++ docs/data/material/getting-started/templates/marketing-page/components/Footer.tsx
