+++ docs/data/material/getting-started/templates/marketing-page/components/Highlights.tsx
