+++ docs/pages/material-ui/getting-started/templates/marketing-page.js
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import AppTheme from 'docs/src/modules/components/AppTheme';
-import LandingPage from 'docs/data/material/getting-started/templates/landing-page/LandingPage';
+import MarketingPage from 'docs/data/material/getting-started/templates/marketing-page/MarketingPage';
 
 export default function Page() {
   return (
     <AppTheme>
-      <LandingPage />
+      <MarketingPage />
     </AppTheme>
   );
 }
