+++ test/regressions/index.js
@@ -68,7 +68,7 @@ const blacklist = [
   'docs-getting-started-templates-sign-in-side/getSignInSideTheme.png', // Theme file
   'docs-getting-started-templates-sign-up/getSignUpTheme.png', // Theme file
   'docs-getting-started-templates-checkout/getCheckoutTheme.png', // Theme file
-  'docs-getting-started-templates-landing-page/getLPTheme.png', // Theme file
+  'docs-getting-started-templates-marketing-page/getMPTheme.png', // Theme file
   'docs-joy-getting-started-templates/TemplateCollection.png', // No public components
   'docs-joy-core-features-automatic-adjustment/ListThemes.png', // No public components
   'docs-joy-tools/PaletteThemeViewer.png', // No need for theme tokens
@@ -179,7 +179,7 @@ const blacklist = [
   'docs-discover-more-languages', // No public components
   'docs-discover-more-showcase', // No public components
   'docs-discover-more-team', // No public components
-  'docs-getting-started-templates-landing-page/LandingPage.png', // Flaky image loading
+  'docs-getting-started-templates-marketing-page/MarketingPage.png', // Flaky image loading
   'docs-getting-started-templates-blog', // Flaky random images
   'docs-getting-started-templates-checkout/AddressForm.png', // Already tested in docs-getting-started-templates-checkout/Checkout
   'docs-getting-started-templates-checkout/PaymentForm.png', // Already tested in docs-getting-started-templates-checkout/Checkout
